moved components from primeqa/pipelines folder to directly primeqa folder
