+++ primeqa/components/indexer/dense.py
@@ -2,7 +2,7 @@ from typing import Union, List
 from dataclasses import dataclass, field
 import json
 
-from primeqa.pipelines.components.base import IndexerComponent
+from primeqa.components.base import IndexerComponent
 from primeqa.ir.dense.colbert_top.colbert.infra.config import ColBERTConfig
 from primeqa.ir.dense.colbert_top.colbert.indexer import Indexer
