+++ primeqa/services/factories.py
@@ -4,16 +4,16 @@ import json
 
 from dataclasses import MISSING
 
-from primeqa.pipelines.components.base import (
+from primeqa.components.base import (
     ReaderComponent,
     RetrieverComponent,
     IndexerComponent,
 )
-from primeqa.pipelines.components.reader.extractive import ExtractiveReader
+from primeqa.components.reader.extractive import ExtractiveReader
 
-from primeqa.pipelines.components.retriever.dense import ColBERTRetriever
+from primeqa.components.retriever.dense import ColBERTRetriever
 
-from primeqa.pipelines.components.indexer.dense import ColBERTIndexer
+from primeqa.components.indexer.dense import ColBERTIndexer
 
 READERS_REGISTRY = {
     ExtractiveReader.__name__: ExtractiveReader,
